requirements.txt
# Document parsing
pymupdf==1.26.3        # Latest stable, manylinux, Python 3.13 support
python-docx==1.1.0

# Embedding
sentence-transformers==2.7.0

# Vector DB (Qdrant + optional Chroma)
qdrant-client==1.15.0  # Latest stable as of July 2025[4]
chromadb==0.4.24       # Only for local dev

# FastAPI backend
fastapi==0.110.1
uvicorn[standard]==0.30.1

# Async operations
aiohttp==3.9.5

# Background processing
httpx==0.27.0

# Data modeling
pydantic

# LLM integration
requests==2.31.0
groq==0.9.0  # Groq API client

# Optional: Docker health checks, linting etc.
python-multipart==0.0.9  # If you're handling file uploads

# For rate limiting and security (optional)
slowapi==0.1.8

# environment management
python-dotenv==1.0.0

